+++ nodemon.json
@@ -1,6 +1,13 @@
 {
-  "watch": ["src","src/ps.js","src/index.html", "src/index.ts", "src/games.ts"],
-  "ext": "ts",
-  "ignore": [""],
-  "exec": "npx ts-node ./src/index.ts"
-}
+    "restartable": "rs",
+    "ignore": [".git", "node_modules/**/node_modules"],
+    "verbose": true,
+    "execMap": { 
+      "ts": "node --require ts-node/register"
+    },
+    "watch": ["src"],
+    "env": {
+      "NODE_ENV": "development"
+    },
+    "ext": "js,json,ts"
+  }
