+++ package.json
@@ -14,7 +14,10 @@
   },
   "devDependencies": {},
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "nodemon src/index.ts",
+    "run": "nodemon src/index.ts",
+    "dev":"nodemon src/index.ts -L"
+
   },
   "repository": {
     "type": "git",
